requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
Levenshtein
thefuzz
jinja2
pymupdf
numpy